+++ R/predict.functions.R
@@ -289,7 +289,7 @@ predict.mbnma <- function(object, times=c(0:max(object$model.arg$jagsdata$time, 
   }
 
   # Convert predicted times to splines
-  if (any(c("rcs", "bs", "ns") %in% object$model.arg$fun)) {
+  if (any(c("rcs", "bs", "ns", "ls") %in% object$model.arg$fun)) {
     spline <- genspline(times, spline=object$model.arg$fun$name, knots=object$model.arg$fun$knots)
   }